refactor(ui): extract team-sync/helpers/invite-panel-dom.ts
Move the DOM helpers that keep the invite/join panel UI in sync into a
single helpers/invite-panel-dom module: applySyncInviteReadinessState,
renderAdminSetupDisclosure, ensureJoinPanelInSetupSection,
setInviteOutputVisibility, setJoinFeedbackVisibility, and
renderInvitePolicySelect. Kept colocated because the open/close cycle
re-renders the select + output. Also deletes the no-op
ensureInvitePanelInAdminSection — the Radix collapsible renders the
panel in place and nothing else calls the helper.
